Add files via upload
These are the 16S rDNA sequences for the manuscript submitted to Frontier's.  Once this is accepted, a copy of the manuscript and the DOI will be added to this manually curated data set.  If you have any questions, contact Feye@uark.edu, Kristina.Feye@usda.gov, or sricke@uark.edu